refactor(broker): extract admin command handlers to broker/admin.rs (#357)
Move handle_admin_cmd() and ADMIN_CMD_NAMES from commands.rs to a
dedicated admin.rs module. Adds 3 new unit tests (permission denied,
kick token invalidation, exit shutdown signal) alongside the 3 moved
tests for a total of 6 admin-specific tests.

Closes #357
