+++ crates/room-cli/src/broker/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod admin;
 pub(crate) mod auth;
 pub(crate) mod commands;
 pub mod daemon;
